requirements.txt
flask
PyPDF2
python-dotenv
openai
waitress